Fix compilation with gcc
```
error: comparison of integer expressions of different signedness: 'uint32_t' and 'int'
```

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6148298
Commit-Queue: Milad Farazmand <mfarazma@redhat.com>
Cr-Commit-Position: refs/heads/main@{#97969}

+++ include/v8-internal.h
@@ -472,7 +472,8 @@ struct TagRange {
     // Need to perform the math with uint32_t. Otherwise, the uint16_ts would
     // be promoted to (signed) int, allowing the compiler to (wrongly) assume
     // that an underflow cannot happen as that would be undefined behavior.
-    return static_cast<uint32_t>(tag) - first <= last - first;
+    return static_cast<uint32_t>(tag) - first <=
+           static_cast<uint32_t>(last) - first;
   }
 
   constexpr bool Contains(TagRange tag_range) const {
